specify parameters required for GitWindow explicitly
AnActionEvent may not have the necessary data in context (for example, when the action is called from "Find Action"). Because of this, trying to create GitWindow directly from the event could result in exception. The default way is to retrieve the data from context in the action class. The action should be disabled when the data is not available.