Use correct api to create new VcsLogUi instance
The purpose of VcsLogUiFactory is to customize how VcsLogManager creates a new VcsLogUi instance. The factory is created specifically for each new VcsLogUi (it has the unique id) and should not be stored, since it has references to the VcsLogManager.

VcsLogUiFactory is not supposed to be used directly by the client code, as VcsLogManager needs to register the newly created VcsLogUi for it to be refreshed properly when new commits appear.