Fix Vcs Log Api Usages in RefactoringHistoryToolbar (#57)
* Use correct api to create new VcsLogUi instance

The purpose of VcsLogUiFactory is to customize how VcsLogManager creates a new VcsLogUi instance. The factory is created specifically for each new VcsLogUi (it has the unique id) and should not be stored, since it has references to the VcsLogManager.

VcsLogUiFactory is not supposed to be used directly by the client code, as VcsLogManager needs to register the newly created VcsLogUi for it to be refreshed properly when new commits appear.

* Wrap VcsLogUi into VcsLogPanel

VcsLogPanel acts as a data provider for standard Vcs Log keys.

* Ensure unique id for newly created VcsLogUi

* Do not keep the last VcsLogUi in the field

As new log tab is created on each click, there is no need to keep the last created tab in the field. Also, because RefactoringHistoryToolbar manages multiple tabs, it should either keep VcsLogUi instances for all tabs or none of them (but not a single one).

* Correctly dispose created log tab

Normally, tabs are opened in the Git tool window and dispose is handled automatically by the VcsLogManager. But since another tool window is used, dispose should be done manually. All opened tabs should be disposed with VcsLogManager, this commit uses ProjectLogListener to do so.