Fix commit refactoring marker by moving it to the separate column
The hack with replacing table renderer for String in order to display labels for commits with detected refactorings does not work since the 2020.3 release. This commit replaces the hack using the new api for custom columns. Refactoring labels are displayed in a separate column which can be enabled from the "Show Columns" menu.